ASoC: msm: update BTSCO rate put function.
The value for the amix command with the numeric strings is
the actual value not the order of the string in the array
of text so the put function has to be updated to handle it.
For example,  the following amix comand:
amix 'Internal BTSCO SampleRatee' 16000'
The value passed from the alsa-lib is not 2, i.e., the 2nd
paramater in the command string:
static const char *btsco_rate_text[] = {"8000", "16000"};
Instead, it is 16000.

CRs-fixed: 364832

Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>
(cherry picked from commit fd8e0b303ba46207f24a3738f8859e84044833fd)

Conflicts:

	sound/soc/msm/mdm9615.c
	sound/soc/msm/msm8930.c
	sound/soc/msm/msm8960.c

Change-Id: I58156f13bec3e4a297d0d4cecc1bf26fb7c7c040
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>