msm: bam_dmux: enable disconnect ack feature support by default
The disconnect ack feature is normally turned on by a handshaking
mechanism between bam_dmux and A2, but on some 9615 variants that
mechanism will not exist.  A2 will still expect the disconnect ack to
be active and might crash if it encounters that it is not active.
Enable the disconnect ack feature by default to prevent this.

CRs-Fixed: 372958 382037
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
(cherry picked from commit b1e7c58c8f7e1495360784929245d30fcd32127e)

Change-Id: I208b55317b8ffe9dc988597b99b855fc9528d341
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>