power: pm8921-charger: Start unplug check worker for DC
It could happen that the device is powered up with a dc charger.
We should start the unplug check worker when the device starts with
dc charger or else we run the risk of not detecting the dc charger
removal.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 90e49c590fccdae53933f7a8fa62a7f0d65c4a0f)

Change-Id: Iaa9e6c56c5766addbf26f1a5052c6ee923ae1013
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>