power: pm8921-charger: disable battery temperature wakeup
The battery temperature interrupts are functional only when there is a
valid charger source present. There is no point making it a wakeup
interrupt, since if the charger is absent the interrupt wont be
triggered (if it is triggered it should be ignored).

Note that for 8921 charging, the software does not take any action if
the battery gets hot/cold. The 8921 hardware state machine stops
charging if the battery temperature goes out of the limits.

For external chargers which depend on the temperature monitoring of
8921, we rely on the fact that the external charger will hold a wakelock
and avoid sleep as long as charging is in progress.

In cases when external charging is finished and the system did go into
suspend, we don't want to wakeup the device only to realize that external
charging was not in progress.

Basically no case requires that we wakeup from battery temperature
interrupts.

(cherry picked from commit a845ec32e6cd85db01c9c5706c0677226e4854b9)

Change-Id: I435e9b4840d5a96d7bca498bee29e49af265dfd3
CRs-Fixed: 395312
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>